+++ inst/sql/sql_server/export/drugera/sqlDrugEraTreemap.sql
@@ -6,9 +6,9 @@ select concept_hierarchy.rxnorm_ingredient_concept_id concept_id,
 	ar1.count_value as num_persons, 
 	1.0*ar1.count_value / denom.count_value as percent_persons,
 	ar2.avg_value as length_of_era
-from (select cast(stratum_1 as int) stratum_1, count_value from @results_database_schema.achilles_results where analysis_id = 900 GROUP BY analysis_id, stratum_1, count_value) ar1
+from (select cast(stratum_1 as bigint) stratum_1, count_value from @results_database_schema.achilles_results where analysis_id = 900 GROUP BY analysis_id, stratum_1, count_value) ar1
 	inner join
-	(select cast(stratum_1 as int) stratum_1, avg_value from @results_database_schema.achilles_results_dist where analysis_id = 907 GROUP BY analysis_id, stratum_1, avg_value) ar2
+	(select cast(stratum_1 as bigint) stratum_1, avg_value from @results_database_schema.achilles_results_dist where analysis_id = 907 GROUP BY analysis_id, stratum_1, avg_value) ar2
 	on ar1.stratum_1 = ar2.stratum_1
 	inner join
 	(
