+++ inst/sql/sql_server/export/procedure/sqlProceduresByType.sql
@@ -4,7 +4,7 @@ select c1.concept_id as procedure_concept_id,
 	c2.concept_name as concept_name, 
 	ar1.count_value as count_value
 from (
-  select cast(stratum_1 as int) stratum_1, cast(stratum_2 as int) stratum_2, count_value
+  select cast(stratum_1 as bigint) stratum_1, cast(stratum_2 as bigint) stratum_2, count_value
   from @results_database_schema.achilles_results
   where analysis_id = 605
   GROUP BY analysis_id, stratum_1, stratum_2, count_value
