+++ inst/sql/sql_server/export/visitdetail/sqlPrevalenceByMonth.sql
@@ -5,8 +5,8 @@ SELECT
 	1000 * (1.0 * num.count_value / denom.count_value) AS y_prevalence_1000pp --prevalence, per 1000 persons
 FROM (
 	SELECT 
-		CAST(stratum_1 AS INT) stratum_1,
-		CAST(stratum_2 AS INT) stratum_2,
+		CAST(stratum_1 AS BIGINT) stratum_1,
+		CAST(stratum_2 AS BIGINT) stratum_2,
 		count_value
 	FROM 
 		@results_database_schema.achilles_results
@@ -20,7 +20,7 @@ FROM (
 	) num
 JOIN (
 	SELECT 
-		CAST(stratum_1 AS INT) stratum_1,
+		CAST(stratum_1 AS BIGINT) stratum_1,
 		count_value
 	FROM 
 		@results_database_schema.achilles_results
